+++ app/src/main/java/ru/frogogo/whitelabel/data/model/ui/home/HomePromotionUiModel.kt
@@ -4,6 +4,7 @@ import android.os.Parcelable
 import androidx.annotation.Keep
 import kotlinx.parcelize.Parcelize
 import ru.frogogo.whitelabel.data.model.ui.PhotoUiModel
+import java.math.BigDecimal
 
 @Keep
 @Parcelize
@@ -11,6 +12,6 @@ data class HomePromotionUiModel(
   val name: String,
   val photo: PhotoUiModel,
   val steps: List<String>,
-  val price: Int,
-  val priceDiscounted: Int,
+  val price: BigDecimal,
+  val priceDiscounted: BigDecimal,
 ) : Parcelable
