+++ app/src/main/java/ru/frogogo/whitelabel/data/model/ui/item/ItemUiModel.kt
@@ -1,13 +1,14 @@
 package ru.frogogo.whitelabel.data.model.ui.item
 
 import ru.frogogo.whitelabel.core.recycler.RecyclerViewItem
+import java.math.BigDecimal
 
 data class ItemUiModel(
   val id: Int,
   val name: String,
   val imageUrl: String,
-  val price: Int,
-  val discountedPrice: Int,
+  val price: BigDecimal,
+  val discountedPrice: BigDecimal,
   val specs: String?,
 ) : RecyclerViewItem {
