+++ app/src/main/java/ru/frogogo/whitelabel/data/model/ui/receipt/ReceiptUiModel.kt
@@ -5,6 +5,7 @@ import androidx.annotation.Keep
 import kotlinx.parcelize.Parcelize
 import ru.frogogo.whitelabel.core.recycler.RecyclerViewItem
 import ru.frogogo.whitelabel.dictionary.ReceiptState
+import java.math.BigDecimal
 import java.util.Date
 
 @Keep
@@ -13,7 +14,7 @@ data class ReceiptUiModel(
   val id: Int,
   val number: Int,
   val date: Date,
-  val value: Int,
+  val value: BigDecimal,
   val state: ReceiptState,
   val shopName: String,
   val rejectReason: ReceiptRejectReasonUiModel?,
